refactor: extract post fetching and filtering logic into PostBrowser component
- Removed complex logic from FetchPost component and created PostBrowser for better separation of concerns.
- Added Spinner component for loading states and improved user experience.
- Integrated existing PostFilters and SentimentBar components into PostBrowser for enhanced functionality.
